Added object configs
