+++ Server/Ship.h
@@ -1,188 +1,49 @@
 #ifndef __SHIP_H__
 #define __SHIP_H__
 
-#ifdef M_SERVER
-#include "../Server/Bubble.h"
-#endif
-#ifdef M_CLIENT
-#include "../robotization dilemma/RenderOut.h"
-#endif
+#include "Sighting.h"
 
-vector<double> intersectPaths(Path* lhs, Path* rhs);
-
-class Movement : public Path { //Order of serialisation
-public:
-  int etype() {
-    return EqnType::EqnTypeApproxable;
-  }
-  float gTimeStamp = 0;        //0
-  fVec3 pos = fVec3(0);        //1
-  //float posUncertainty = 0;
-  fVec3 vel = fVec3(0);        //2
-  //float velUncertainty = 0;
-  fVec3 acc = fVec3(0);        //3
-  //float accUncertainty = 0;
-  int type;                    //4
-  string pathData;             //5
-  double radius;               //6
-  Eqnsys getEquations(bool b) {//approximate
-    Eqnsys res;
-    if (!b) { //Parametric curve
-      Eqnsys resh;
-
-      resh.eqns['a'] = Equation<double>({ { acc.x / 2.0f, "tt" },{ vel.x, "t" } ,{ pos.x, "" },{ -1, "a" } });
-      resh.eqns['b'] = Equation<double>({ { acc.y / 2.0f, "tt" },{ vel.y, "t" } ,{ pos.y, "" },{ -1, "b" } });
-      resh.eqns['c'] = Equation<double>({ { acc.z / 2.0f, "tt" },{ vel.z, "t" } ,{ pos.z, "" },{ -1, "c" } });
-     
-
-      res.eqns['t'] =
-        Equation<double>({ { 1, "x" },{ -1, "a" } }) * Equation<double>({ { 1, "x" },{ -1, "a" } }) +
-        Equation<double>({ { 1, "y" },{ -1, "b" } }) * Equation<double>({ { 1, "y" },{ -1, "b" } }) +
-        Equation<double>({ { 1, "z" },{ -1, "c" } }) * Equation<double>({ { 1, "z" },{ -1, "c" } }) +
-        Equation<double>({ { -radius*radius, "" } });
-      res.eqns['t'].substitute(resh.eqns['a'], 'a');
-      res.eqns['t'].substitute(resh.eqns['b'], 'b');
-      res.eqns['t'].substitute(resh.eqns['c'], 'c');
-    }
-    else { //Moving sphere
-      res.eqns['x'] = Equation<double>({ { acc.x / 2.0f, "tt" },{ vel.x, "t" } ,{ pos.x, "" },{ -1, "x" } });
-      res.eqns['y'] = Equation<double>({ { acc.y / 2.0f, "tt" },{ vel.y, "t" } ,{ pos.y, "" },{ -1, "y" } });
-      res.eqns['z'] = Equation<double>({ { acc.z / 2.0f, "tt" },{ vel.z, "t" } ,{ pos.z, "" },{ -1, "z" } });
-    }
-    return res;
-  }
-  /*float reachMaxVelIn(float maxVelocity, bool& will) {
-    //-1: No acceleration, cant be solved.
-    //0 .. 2: number of solutions
-    int numOfSolut;
-    float sol[2];
-    float a = pow(acc.length(), 2);
-    float b = 2 * (vel.x*acc.x + vel.y*acc.y + vel.z*acc.z);
-    float c = pow(vel.length(), 2) - pow(maxVelocity, 2); //solve (a+vt)^2 = vmax^2 for t
-    if (a == 0)
-    {
-      will = false;
-      return 0;
-    } else {
-      will = true;
-      solve2(a, b, c, sol, numOfSolut); //calculate time of reacing maximum velocity
-      if(numOfSolut < 2) {
-        throw 1;
-        return 0;
-      } else {
-        return max(sol[0], sol[1]);
-      }
-    }
-    throw 1;
-    return 0;
-  }*/
-  Movement goTo(float gTime, float maxVelocity);
-  /*Movement goForwardTo(float time, float maxVelocity) {
-    Movement res;
-    res.timestamp = time;
-    float dt = time - timestamp;
-    bool will = false;
-    float reachMaxT = reachMaxVelIn(maxVelocity, will);
-    if (!will) {
-      res.pos = pos + vel * dt;
-      res.vel = vel;
-      res.acc = 0;
-    }
-    else
-    {
-      if (reachMaxT < 0)
-      {
-        throw 1;
-      }
-      if (reachMaxT > dt)
-      {
-        res.pos = pos + vel * dt + acc*(pow(dt, 2) / 2);
-        res.vel = vel + acc*dt;
-        res.acc = acc;
-      }
-      else
-      {
-        fVec3 startVelocity = vel;
-        res.vel = vel + acc*reachMaxT;
-        res.pos = pos + startVelocity*dt + vel*(dt - reachMaxT) + acc*(pow(reachMaxT, 2) / 2);
-        res.acc = 0;
-      }
-    }
-    //timestamp = time;
-  }
-  Movement goBackTo(float time, float maxVelocity) {
-    Movement res;
-    res.timestamp = time;
-    float dt = time - timestamp;
-    res.pos = pos + vel * dt + acc*(pow(dt, 2) / 2);
-    res.vel = vel + acc*dt;
-    res.acc = acc;
-  }*/
-
-  void get(unsigned char** data, int &DataLen);
-  void set(unsigned char* data, int DataLen);
-
-  ~Movement();
-};
-
-class Sighting {
-public:
-  vector<Movement*> keyframes; //time - keyframe. Sorted.
-  int getLastSmaller(float t);
-  Movement estimatePos(float t, float maxVelocity);
-  
-#ifdef M_CLIENT
-  void drawSighting(float camcx, float camcy, float camcz, float d, float maxVel);
-#endif
-
-  void getSighting(unsigned char** data, int &DataLen);
-  void setSighting(unsigned char* data, int DataLen);
-
-  void clearKeyframes();
-  ~Sighting();
-};
+extern class Object;
+extern Object* selected;
 
 class Drone;
 
 class Object {       //Order of serialisation
+protected:
+  fVec3 _relativePos; //0
+  int _maxHealth;     //2
+  int _health;        //3
+  float _radius;      //4
 public:
+
+  Object(fVec3 relativePos, int maxHealth, float radius, int health) {
+    _relativePos = relativePos;
+    _maxHealth = maxHealth;
+    _health = health;
+    _radius = radius;
+  }
+
   Drone* parentShip;
-  fVec3 relativePos; //0
+
   static enum Type {
     Ship = 1,
     Shield = 2,
     Sensor = 3,
     Computer = 4,
     Generator = 5,
-    Storage = 6
+    Storage = 6,
+    Engine = 7
   };
-  int type;         //1
-  int maxHealth;    //2
-  int health;       //3
-  float radius;     //4
-  float energy;     //5
-  float maxEnergy;  //6
+  virtual int type() {throw 1; return 0;}         //1
+  virtual float usedEnergy() {throw 1; return 0;}
+  virtual float prodEnergy() {return 0;}
 
   Movement getMovement();
 
 #ifdef M_CLIENT
   list< pair<double, pair<Object*, Path*>>> getIntersect(vec3<double> ori, vec3<double> dir);
-  void setSidebar() {
-    //cout << type << " clicked " << health/float(maxHealth) << endl;
-    if(type == Sensor) {
-      Graphics::setElements(reinterpret_cast<Graphics::PanelHwnd>(Graphics::getElementById("objectIngameMenuSidebar")), "html/sensor_settings.html");
-      reinterpret_cast<Graphics::LabelHwnd>(Graphics::getElementById("objectSensorSidebarHealth"))->text = "Health: " + to_string(health) + " / " + to_string(maxHealth);
-      reinterpret_cast<Graphics::LabelHwnd>(Graphics::getElementById("objectSensorSidebarEnergyLabel"))->text = " / " + to_string(maxEnergy);
-      return;
-    }
-    if (type == Generator) {
-      Graphics::setElements(reinterpret_cast<Graphics::PanelHwnd>(Graphics::getElementById("objectIngameMenuSidebar")), "html/generator_settings.html");
-      reinterpret_cast<Graphics::LabelHwnd>(Graphics::getElementById("objectGeneratorSidebarHealth"))->text = "Health: " + to_string(health) + " / " + to_string(maxHealth);
-      reinterpret_cast<Graphics::LabelHwnd>(Graphics::getElementById("objectGeneratorSidebarMaxEnergyLabel"))->text = "Max output: " + to_string(maxEnergy);
-      return;
-    }
-    cout << "Unimplemented type " << type << endl;
-  }
+  virtual void setSidebarElement();
+  virtual void setSidebar();
   void drawObject(float camcx, float camcy, float camcz, float d);
 #endif
 
@@ -201,38 +62,101 @@ public:
   ~Object();
 };
 
-class Drone : public Object {
+class Sensor : public Object {       //Order of serialisation
+private:
+  float _energy;
+  float _maxEnergy;
+public:
+  Sensor(fVec3 relativePos, int maxHealth, float radius, int health, float maxEnergy) : Object(relativePos, maxHealth, radius, health) {
+    _energy = 0;
+    _maxEnergy = maxEnergy;
+  }
+
+  int type() {return Type::Sensor;}
+
+  float usedEnergy() { return _energy; }
+
+  void setEnergy(float val) {
+    _energy = max(min(val, _maxEnergy), 0.0f);
+  }
+
+#ifdef M_CLIENT
+  void setSidebarElement();
+  void setSidebar();
+#endif
+
+  void getStatus(unsigned char** data, int &DataLen);
+  void setStatus(unsigned char* data, int DataLen);
+};
+
+class Engine : public Object {       //Order of serialisation
+private:
+  float _maxEnergy;
+  fVec3 _accel;
+public:
+  Engine(fVec3 relativePos, int maxHealth, float radius, int health, float maxEnergy, fVec3 accel) : Object(relativePos, maxHealth, radius, health) {
+    _accel = accel;
+    _maxEnergy = maxEnergy;
+  }
+
+  int type() { return Type::Engine; }
+  float usedEnergy() { return _accel.sqrlen(); }
+
+  void setComponent(int c, float val) {
+    _accel[c] = val;
+  }
+
+#ifdef M_CLIENT
+  void setSidebarElement();
+  void setSidebar();
+#endif
+
+  void getStatus(unsigned char** data, int &DataLen);
+  void setStatus(unsigned char* data, int DataLen);
+};
+
+class Generator : public Object {       //Order of serialisation
+private:
+  float _maxEnergy;
+public:
+  Generator(fVec3 relativePos, int maxHealth, float radius, int health, float maxEnergy) : Object(relativePos, maxHealth, radius, health) {
+    _maxEnergy = maxEnergy;
+  }
+  int type() { return Type::Generator; }
+  float usedEnergy() { return 0; }
+  float prodEnergy() { return _maxEnergy; }
+
+#ifdef M_CLIENT
+  void setSidebarElement();
+  void setSidebar();
+#endif
+
+  void getStatus(unsigned char** data, int &DataLen);
+  void setStatus(unsigned char* data, int DataLen);
+};
+
+class Drone {
 public:
   Movement mov;
 };
 
 class Ship : public Drone {
-public:
-  fVec3 accel = fVec3(0);
-
+private:
   bool canMove = false; //is the player open to moving / are we waiting for a move.
-
+public:
   Ship() {
-    Object* no = new Object();
+    Object* no = new ::Generator({ 100,0,0 }, 1000, 100, 800, 100000);
     no->parentShip = this;
-    no->relativePos = {100,0,0};
-    no->type = Generator;
-    no->maxHealth = 1000;
-    no->health = 800;
-    no->radius = 100;
-    no->maxEnergy = 100000;
 
     objects.push_back(no);
 
-    no = new Object();
+    no = new ::Sensor({ -100,0,0 }, 1000, 100, 600, 100000);
     no->parentShip = this;
-    no->relativePos = { -100,0,0 };
-    no->type = Sensor;
-    no->maxHealth = 1000;
-    no->health = 600;
-    no->radius = 100;
-    no->maxEnergy = 100000;
-    no->energy = 0;
+
+    objects.push_back(no);
+
+    no = new ::Engine({ 0,173.2f ,0 }, 1000, 100, 900, 100000, {0, 0, 0});
+    no->parentShip = this;
 
     objects.push_back(no);
   }
@@ -263,6 +187,9 @@ public:
 
   void newTurn(int id);
 
+  void setSidebarElement();
+  void setSidebar();
+
   void setSidebar(vec3<double> ori, vec3<double> dir) {
     list< pair<double, pair<Object*, Path*>>> inters;
 
@@ -278,6 +205,8 @@ public:
 
      if(inters.size()) {
        inters.begin()->second.first->setSidebar();
+     } else {
+       setSidebar();
      }
   }
 
@@ -286,15 +215,16 @@ public:
 
   bool packetRecv(unsigned char *Data, int Id, int DataLen, NetworkC* thisptr);
 #endif
-  Object* getObject(int type);
+  /*Object* getObject(int type);
   Object* getGenerator();
   Object* getSensor();
   void  setSensorEnergy(float energy);
   float getSensorEnergy();
-  float getMaximumSensorEnergy();
+  float getMaximumSensorEnergy();*/
 
   float getTotalShipEnergy();
   float getRemainingShipEnergy();
+  float getSpentShipEnergy();
 
   void getStatus(unsigned char** data, int &DataLen);
   void setStatus(unsigned char* data, int DataLen);
@@ -308,79 +238,9 @@ public:
   ~Ship();
 };
 
-//void shipPacketRecv(unsigned char *Data, int Id, int DataLen, NetworkS* thisptr, Ship* ship);
-
+#ifdef M_CLIENT
+//extern lis<Command> commands
 extern Ship* ship;
-
-/*
-struct Command
-{
-  public:
-  fVec3 accel = fVec3(0);
-  bool didFire = false;
-  unsigned int aim;
-  int sensorEnergy = 0;
-  friend std::ostream& operator<<(std::ostream& os, const Command& c);
-};
-
-struct SensorDataElement
-{
-  enum Type {Active, Passive};
-  Type detectType;
-  unsigned int player;
-  fVec3 place;
-  fVec3 velocity;
-  fVec3 nextPlace;
-  bool sureFire;
-  friend std::ostream& operator<<(std::ostream& os, const SensorDataElement& sde);
-};
-
-class Object
-{
-protected:
-  fVec3 place;
-  fVec3 velocity;
-  int maxVelocity;
-
-public:
-  Object(fVec3 place, fVec3 velocity, int maxVelocity) : 
-    place(place), velocity(velocity), maxVelocity(maxVelocity) {}
-  void move(fVec3 accel, float time);
-};
-
-class Ship : public Object
-{
-private:
-  const unsigned int owner;
-  int maxAcceleration = INITIAL_MAX_ACC;
-  int maxSensorEnergy = INITIAL_MAX_SENSORENERGY;
-  int maxGeneratorEnergy = INITIAL_MAX_GENERATORENERGY;
-  int hullRadius = INITIAL_HULL_RADIUS;
-  bool destroyed = false;
-  Command command;
-  std::vector<SensorDataElement> sensorData;
-  int getSpentEnergy() const;
-
-public:
-  Ship(fVec3 place, int owner) : 
-    Object(place, place * ((-1)*INITIAL_VELOCITY/place.length()), INITIAL_MAX_VELOCITY),
-    owner(owner) {}
-  void getCommand();
-  void flushSensorData();
-  void sense(SensorDataElement sde);
-  void giveSensorData() const;
-  void move(float time);
-  void destroy();
-  int getAim() const;
-  fVec3 getPlace() const;
-  fVec3 getVelocity() const;
-  int getHullRadius() const;
-  bool isDestroyed() const;
-  int getSensorRadiation() const;
-  int getVisibility() const;
-  bool didFire() const;
-  unsigned int getOwner() const;
-  friend std::ostream& operator<<(std::ostream& os, const Ship& s);
-};*/
+#endif
 
 #endif
