+++ robotization dilemma/html/engine_settings.html
@@ -1,0 +1,140 @@
+<body>
+	<label
+		id = "objectEngineSidebarName"
+		minrelx = "0.0"
+		minrely = "1.0"
+		minabsx = "5.0"
+		minabsy = "-45.0"
+		maxrelx = "1.0"
+		maxrely = "1.0"
+		maxabsx = "-5.0"
+		maxabsy = "-5.0"
+		bgcolor = "ff070918"
+		activecolor = "ff190077"
+		textcolor = "ff838fa1"
+		align = "1"
+	>Engine</label>
+	<label
+		id = "objectEngineSidebarHealth"
+		minrelx = "0.0"
+		minrely = "1.0"
+		minabsx = "5.0"
+		minabsy = "-95.0"
+		maxrelx = "1.0"
+		maxrely = "1.0"
+		maxabsx = "-5.0"
+		maxabsy = "-55.0"
+		bgcolor = "ff070918"
+		activecolor = "ff190077"
+		textcolor = "ff838fa1"
+		align = "1"
+	>0 / 0</label>
+	<label
+		id = "objectEngineSidebarEnergyLabel"
+		minrelx = "0.0"
+		minrely = "1.0"
+		minabsx = "5.0"
+		minabsy = "-145.0"
+		maxrelx = "1.0"
+		maxrely = "1.0"
+		maxabsx = "-5.0"
+		maxabsy = "-105.0"
+		bgcolor = "ff070918"
+		activecolor = "ff190077"
+		textcolor = "ff838fa1"
+		align = "1"
+	>0 / 0</label>
+	<label
+		id = "objectEngineSidebarAccInputXLabel"
+		minrelx = "0.0"
+		minrely = "1.0"
+		minabsx = "5.0"
+		minabsy = "-195.0"
+		maxrelx = "0.33"
+		maxrely = "1.0"
+		maxabsx = "-5.0"
+		maxabsy = "-155.0"
+		bgcolor = "ff070918"
+		activecolor = "ff190077"
+		textcolor = "ff838fa1"
+		align = "1"
+	>x</label>
+	<label
+		id = "objectEngineSidebarAccInputYLabel"
+		minrelx = "0.33"
+		minrely = "1.0"
+		minabsx = "5.0"
+		minabsy = "-195.0"
+		maxrelx = "0.67"
+		maxrely = "1.0"
+		maxabsx = "-5.0"
+		maxabsy = "-155.0"
+		bgcolor = "ff070918"
+		activecolor = "ff190077"
+		textcolor = "ff838fa1"
+		align = "1"
+	>y</label>
+	<label
+		id = "objectEngineSidebarAccInputZLabel"
+		minrelx = "0.67"
+		minrely = "1.0"
+		minabsx = "5.0"
+		minabsy = "-195.0"
+		maxrelx = "1.0"
+		maxrely = "1.0"
+		maxabsx = "-5.0"
+		maxabsy = "-155.0"
+		bgcolor = "ff070918"
+		activecolor = "ff190077"
+		textcolor = "ff838fa1"
+		align = "1"
+	>z</label>
+	<input
+		id = "objectEngineSidebarAccInputX"
+		minrelx = "0.0"
+		minrely = "1.0"
+		minabsx = "5.0"
+		minabsy = "-245.0"
+		maxrelx = "0.33"
+		maxrely = "1.0"
+		maxabsx = "-5.0"
+		maxabsy = "-205.0"
+		bgcolor = "ff070918"
+		activecolor = "ff190077"
+		textcolor = "ff838fa1"
+		inputfunc = "ingameEngineSidebarAccInputX"
+		validatorfunc = "floatValidator"
+	>0</input>
+	<input
+		id = "objectEngineSidebarAccInputY"
+		minrelx = "0.33"
+		minrely = "1.0"
+		minabsx = "5.0"
+		minabsy = "-245.0"
+		maxrelx = "0.67"
+		maxrely = "1.0"
+		maxabsx = "-5.0"
+		maxabsy = "-205.0"
+		bgcolor = "ff070918"
+		activecolor = "ff190077"
+		textcolor = "ff838fa1"
+		inputfunc = "ingameEngineSidebarAccInputY"
+		validatorfunc = "floatValidator"
+	>0</input>
+	<input
+		id = "objectEngineSidebarAccInputZ"
+		minrelx = "0.67"
+		minrely = "1.0"
+		minabsx = "5.0"
+		minabsy = "-245.0"
+		maxrelx = "1.0"
+		maxrely = "1.0"
+		maxabsx = "-5.0"
+		maxabsy = "-205.0"
+		bgcolor = "ff070918"
+		activecolor = "ff190077"
+		textcolor = "ff838fa1"
+		inputfunc = "ingameEngineSidebarAccInputZ"
+		validatorfunc = "floatValidator"
+	>0</input>
+</body>
