+++ robotization dilemma/html/game_screen.html
@@ -37,10 +37,10 @@
 	>
 		<button
 			id = "objectMainGameExitButton"
-			minrelx = "0.0"
-			minrely = "0.8"
+			minrelx = "0.5"
+			minrely = "1.0"
 			minabsx = "5.0"
-			minabsy = "5.0"
+			minabsy = "-45.0"
 			maxrelx = "1.0"
 			maxrely = "1.0"
 			maxabsx = "-5.0"
@@ -50,29 +50,14 @@
 			textcolor = "ff838fa1"
 			callback = "ingameMenuExitButton"
 		>Exit to menu</button>
-		<label
-			id = "objectMainGameTurnLabel"
-			minrelx = "0.0"
-			minrely = "0.6"
-			minabsx = "5.0"
-			minabsy = "5.0"
-			maxrelx = "0.5"
-			maxrely = "0.8"
-			maxabsx = "-5.0"
-			maxabsy = "-5.0"
-			bgcolor = "ff070918"
-			activecolor = "ff190077"
-			textcolor = "ff838fa1"
-			align = "1"
-		>Not your turn</label>
 		<button
 			id = "objectMainGameCommitButton"
-			minrelx = "0.5"
-			minrely = "0.6"
+			minrelx = "0.0"
+			minrely = "1.0"
 			minabsx = "5.0"
-			minabsy = "5.0"
-			maxrelx = "1.0"
-			maxrely = "0.8"
+			minabsy = "-45.0"
+			maxrelx = "0.5"
+			maxrely = "1.0"
 			maxabsx = "-5.0"
 			maxabsy = "-5.0"
 			bgcolor = "ff070918"
@@ -87,9 +72,9 @@
 			minabsx = "0.0"
 			minabsy = "0.0"
 			maxrelx = "1.0"
-			maxrely = "0.6"
+			maxrely = "1.0"
 			maxabsx = "0.0"
-			maxabsy = "0.0"
+			maxabsy = "-50.0"
 			bgcolor = "ff070918">
 		</div>
 	</div>
