+++ robotization dilemma/html/sensor_settings.html
@@ -2,9 +2,9 @@
 	<label
 		id = "objectSensorSidebarName"
 		minrelx = "0.0"
-		minrely = "0.66"
+		minrely = "1.0"
 		minabsx = "5.0"
-		minabsy = "5.0"
+		minabsy = "-45.0"
 		maxrelx = "1.0"
 		maxrely = "1.0"
 		maxabsx = "-5.0"
@@ -17,13 +17,13 @@
 	<label
 		id = "objectSensorSidebarHealth"
 		minrelx = "0.0"
-		minrely = "0.33"
+		minrely = "1.0"
 		minabsx = "5.0"
-		minabsy = "5.0"
+		minabsy = "-95.0"
 		maxrelx = "1.0"
-		maxrely = "0.66"
+		maxrely = "1.0"
 		maxabsx = "-5.0"
-		maxabsy = "-5.0"
+		maxabsy = "-55.0"
 		bgcolor = "ff070918"
 		activecolor = "ff190077"
 		textcolor = "ff838fa1"
@@ -32,29 +32,29 @@
 	<input
 		id = "objectSensorSidebarEnergyInput"
 		minrelx = "0.0"
-		minrely = "0.0"
+		minrely = "1.0"
 		minabsx = "5.0"
-		minabsy = "5.0"
+		minabsy = "-145.0"
 		maxrelx = "0.5"
-		maxrely = "0.33"
+		maxrely = "1.0"
 		maxabsx = "-5.0"
-		maxabsy = "-5.0"
+		maxabsy = "-105.0"
 		bgcolor = "ff070918"
 		activecolor = "ff190077"
 		textcolor = "ff838fa1"
 		inputfunc = "ingameSensorSidebarEnergyInput"
-		validatorfunc = "numericalValidator"
+		validatorfunc = "floatValidator"
 	>0</input>
 	<label
 		id = "objectSensorSidebarEnergyLabel"
 		minrelx = "0.5"
-		minrely = "0.0"
+		minrely = "1.0"
 		minabsx = "5.0"
-		minabsy = "5.0"
+		minabsy = "-145.0"
 		maxrelx = "1.0"
-		maxrely = "0.33"
+		maxrely = "1.0"
 		maxabsx = "-5.0"
-		maxabsy = "-5.0"
+		maxabsy = "-105.0"
 		bgcolor = "ff070918"
 		activecolor = "ff190077"
 		textcolor = "ff838fa1"
