+++ robotization dilemma/html/ship_settings.html
@@ -1,6 +1,6 @@
 <body>
 	<label
-		id = "objectGeneratorSidebarName"
+		id = "objectShipSidebarName"
 		minrelx = "0.0"
 		minrely = "1.0"
 		minabsx = "5.0"
@@ -13,9 +13,9 @@
 		activecolor = "ff190077"
 		textcolor = "ff838fa1"
 		align = "1"
-	>Genetor</label>
+	>Ship</label>
 	<label
-		id = "objectGeneratorSidebarHealth"
+		id = "objectShipSidebarEnergyLabel"
 		minrelx = "0.0"
 		minrely = "1.0"
 		minabsx = "5.0"
@@ -29,19 +29,4 @@
 		textcolor = "ff838fa1"
 		align = "1"
 	>0 / 0</label>
-	<label
-		id = "objectGeneratorSidebarMaxEnergyLabel"
-		minrelx = "0.0"
-		minrely = "1.0"
-		minabsx = "5.0"
-		minabsy = "-145.0"
-		maxrelx = "1.0"
-		maxrely = "1.0"
-		maxabsx = "-5.0"
-		maxabsy = "-105.0"
-		bgcolor = "ff070918"
-		activecolor = "ff190077"
-		textcolor = "ff838fa1"
-		align = "1"
-	>0</label>
 </body>
