+++ robotization dilemma/robotization dilemma.vcxproj.filters
@@ -72,6 +72,9 @@
     <ClCompile Include="RenderOut.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="..\Server\Sighting.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="Graphics.h">
